+++ CHANGELOG.md
@@ -23,6 +23,7 @@ Replacement `tsi1` indexes will be automatically generated on startup for shards
 1. [20403](https://github.com/influxdata/influxdb/pull/20403): Improve error message shown when `influx` CLI can't find an org by name.
 1. [20411](https://github.com/influxdata/influxdb/pull/20411): Fix logging initialization for storage engine.
 1. [20456](https://github.com/influxdata/influxdb/pull/20456): Automatically build `tsi1` indexes for shards that need it instead of falling back to `inmem`.
+1. [20442](https://github.com/influxdata/influxdb/pull/20442): Don't return 500 codes for partial write failures.
 
 ## v2.0.3 [2020-12-14]
 ----------------------